frontend/requirements.txt

streamlit==1.28.0
plotly==6.5.1
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0